Strict compiler options

+++ 0088.cpp
@@ -70,6 +70,6 @@ int main()
 	assert(b.size() == 10);
 	Solution s;
 	s.merge(a, size_a, b, size_b);
-	for (int i = 0; i < a.size(); ++i)
+	for (size_t i = 0; i < a.size(); ++i)
 		std::cout << a[i] << " ";
 }
